factor common export code into export.js
